refactor: rename achievements handler
